requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Development
black==23.10.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0 